Updated server.js

+++ server/server.js
@@ -55,9 +55,20 @@ app.get('/todos/:id', (req, res) => {
 
 app.delete('/todos/:id', (request, response) => {
 	var id = req.params.id;
-	Todo.findByIdAndRemove(id).then((todo) => {
-		response.status(201).send(todo);
-	}).catch((err) => response.status(404).send());
+
+  if (!ObjectID.isValid(id)) {
+    return res.status(404).send();
+  }
+
+  Todo.findByIdAndRemove(id).then((todo) => {
+    if (!todo) {
+      return res.status(404).send();
+    }
+
+    res.send({todo});
+  }).catch((e) => {
+    res.status(400).send();
+  });
 });
 
 app.listen(port, () => {
